Made minor repairs to ffi1001
Improved ffi1001 representation of dates to include microsecond.  Microsecond is not included in strftime until python 2.6, so I implemented a similar methodology

Improved ffi1001 representation of fill_value and scale.  These were previously stored as strings, but make more sense as float or integer values.  For consistency, TFLAG also has a fill value.  This is not strictly speaking correct.  TFLAG is a string and a factor would actually be a character repeat.  This may be removed