Allowing for typed values keyword to PNFVariable
PseudoNetCDFVariables have long accepted a values keyword to supply values on creation, but the meta data of the values class were potentially over-written.  The values variable had its __dict__ over-written, instead of updated.  The problem is evident in MaskedArrays.  The _data and _mask variables were being lost because they were in __dict__.  I have updated Variable object to update __dict__ when it exists

git-svn-id: https://dawes.sph.unc.edu/uncaqmlsvn/PseudoNetCDF/trunk@350 ea03d40b-b492-4883-a61c-0e53214263a1