Fixing sphere definition and adding ease of use functions.
rsphere must be passed as a and b to pyproj.Proj

adding getcdo function to improve ease of use. Will likely lead to an
--output-format cdogrid

adding getlonlatcoordstr function to improve ease of use. Will likely
lead to an extract option that references another file.

Should we improve extract to use great circle calculations or is
lat/lon close enough in these purposes.