Adding aerosol_names options including named mech options
This improves hardcoded treatment of aerosol units in the UAMIV memmap
and reader classes. First, the default aerosol species have been
updated to be inclusive of v6.3. Second, file instances can optionally
specify their chemparam. With a specified chemparam, the aerosols will
be specific to that implementation. The chemparam options can be seen
with the following command.

pydoc PseudoNetCDF.camxfiles.aerosol_names