Adding generalized reader
pncopen uses the format keyword to select a reader and args/kwds to actually open it. This provides a generalized reader necessary for xarray datastore. Previously this was called anyfile and could not specify a reader, but it is backward compatible.

The isMine method of aermod had to be fixed to support anyfile.