Adding Wrap class
The Wrap Class allows pseudonetcdf to modify a netCDF file without
loading all of the data into memory first. It does this by acting
as a pass through for attributes, variables, and dimensions.

The only missing functionality is to allow writing meta-data to
netCDF variables when in write-only mode. This is useful for
presenting adding conventions to a file.