Units handling for ICARTT independent variable
In version 2 of the ICARTT file format standards,
the independent variable is specified as:
Variable short name, variable unit, variable standard name, [optional variable long name]

In version 1, the defining the units alone was permitted.

For lines that split on ',' (i.e. version 2)
the variable name and units are both used.

Signed-off-by: Christopher Holmes <cdholmes@fsu.edu>