grib2 reader works
This has only been tested with NOAA global GDAS forecasts.
It needs more testing with a variety of inputs, because
the current features are largely heuristic.

Supports:
  * multiple z-coordinates
  * untested, multiple times
  * multiple variables in each z-coordinate

Variables may share z-coordinates. In some cases, they are not
reported on all the same z-coordinates. Missing values will be
treated as nan for now, though a masked variable would be better.