Distinguishing bewteen levels and integrals
Some variables are reported on levels, while others report integrals 3000-0
for example. Previously, these were being lumped together as a single vertical
coordinate (1500, so 0, 100, 1500). This is misleading. Now, vertical coordinates
include variants. Integrals and levels. The variables associated with them include
the full name of the vertical coordinate.

This is necessary because some variables are present on the integrals and the levels