Time-independent Functionality
The ioapi getTimes function failed when SDATE was set to -635,
which is used sometimes for time independent data. This had
cascading effects on all functions that used getTimes as part
of the meta-data refresh.

This bug-fix essentially updates the file to have a 1970001
SDATE and warns the user. This is far preferable to the previous
work arounds.