Adding Audit
It is easy to get IOAPI metadata wrong. When you do, it will
still work in netcdf mode, but will fail when being read by
IOAPI. This update adds an audit system. The audit system will
tell you what is wrong, but will not update it. This is useful
for checking products during development to ensure that they
will be CMAQ-ready when they are complete.